requirements.txt
Pillow>=10.0.0
rembg>=2.0.50
numpy>=1.24.0
pytest>=7.4.0
